feat:리뷰 작성 이미지 등록 추가

+++ cleanBooth/src/main/java/cleanBooth/cleanBooth/Item/Dto/PostReviewDto.java
@@ -10,11 +10,11 @@ import java.util.List;
 public class PostReviewDto {
     private String goodDescription;
     private String badDescription;
-    private Float score;
+    private Double score;
     private List<String> photos;
 
     @Builder
-    public PostReviewDto(String goodDescription, String badDescription, Float score, List<String> image){
+    public PostReviewDto(String goodDescription, String badDescription, Double score, List<String> image){
         this.goodDescription = goodDescription;
         this.badDescription = badDescription;
         this.score = score;
